Updated README.md to include Assignment 4 thoughts.
As I (Ken Nelson) have stated in the past, I cannot stand python as a language. The tools are horrible, pip being a prime example. That said
we used to pip to assist in creating the requirements.txt file, although we are not certain that it accurately reflected the projects
dependencies. We had to manually enter a number of dependencies that the pip freeze command missed. We also added a dependencies.txt file
to list all the dependency versions, although we feel that is inadequate as well. In reading some online forums, pip freeze is not seen
as a good solution to use (read: https://medium.com/@tomagee/pip-freeze-requirements-txt-considered-harmful-f0bce66cf895).

Co-authored-by: Ken Nelson <kenneth.nelson@cox.net>
Co-authored-by: Dianna Mega <dmega@asu.edu>